requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
face_recognition==1.3.0
faiss-cpu==1.7.1
Flask==2.0.1
tqdm==4.62.3